requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
Authlib==1.3.2
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
multidict==6.1.0
passlib==1.7.4
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
securely==0.2.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.44.0
typing_extensions==4.12.2
uvicorn==0.33.0
yarl==1.15.2
